[#11348] feat(flink-connector): Flink connector view support to Iceberg and Paimon catalogs
- Fix GravitinoPaimonCatalog.dropTable to fall back to dropView when purgeTable returns false
- Move 9 view ITs from FlinkHiveCatalogIT into shared FlinkCommonIT with supportViewOperation hook
- Enable view tests for Iceberg (Hive+REST) and Paimon (Hive+JDBC) backends
