+++ flink-connector/flink-common/src/test/java/org/apache/gravitino/flink/connector/integration/test/hive/FlinkHiveCatalogIT.java
@@ -60,7 +60,6 @@ import org.apache.flink.table.catalog.hive.factories.HiveCatalogFactoryOptions;
 import org.apache.flink.table.catalog.hive.util.Constants;
 import org.apache.flink.types.Row;
 import org.apache.gravitino.NameIdentifier;
-import org.apache.gravitino.Namespace;
 import org.apache.gravitino.catalog.hive.HiveConstants;
 import org.apache.gravitino.catalog.hive.HiveStorageConstants;
 import org.apache.gravitino.flink.connector.CatalogPropertiesConverter;
@@ -72,10 +71,7 @@ import org.apache.gravitino.flink.connector.store.GravitinoCatalogStoreFactoryOp
 import org.apache.gravitino.integration.test.container.ContainerSuite;
 import org.apache.gravitino.integration.test.util.TestDatabaseName;
 import org.apache.gravitino.rel.Column;
-import org.apache.gravitino.rel.SQLRepresentation;
 import org.apache.gravitino.rel.Table;
-import org.apache.gravitino.rel.View;
-import org.apache.gravitino.rel.ViewCatalog;
 import org.apache.gravitino.rel.expressions.transforms.Transform;
 import org.apache.gravitino.rel.expressions.transforms.Transforms;
 import org.apache.gravitino.rel.types.Types;
@@ -749,275 +745,14 @@ public abstract class FlinkHiveCatalogIT extends FlinkCommonIT {
         true);
   }
 
-  @Test
-  public void testCreateView() {
-    String schemaName = "test_hive_create_view_db";
-    String viewName = "test_view_create";
-    String tableName = "test_view_base_table";
-    doWithSchema(
-        currentCatalog(),
-        schemaName,
-        catalog -> {
-          TestUtils.assertTableResult(
-              sql("CREATE TABLE %s (id INT, name STRING) WITH ('connector'='hive')", tableName),
-              ResultKind.SUCCESS);
-          TestUtils.assertTableResult(
-              sql(
-                  "CREATE VIEW %s COMMENT 'view comment' AS SELECT id, name FROM %s",
-                  viewName, tableName),
-              ResultKind.SUCCESS);
-
-          // Verify via Gravitino ViewCatalog
-          ViewCatalog viewCatalog = catalog.asViewCatalog();
-          View view = viewCatalog.loadView(NameIdentifier.of(schemaName, viewName));
-          Assertions.assertEquals(viewName, view.name());
-          Assertions.assertEquals("view comment", view.comment());
-          Assertions.assertEquals(1, view.representations().length);
-          Assertions.assertInstanceOf(SQLRepresentation.class, view.representations()[0]);
-
-          // Verify via Flink catalog API
-          Optional<Catalog> flinkCatalog = tableEnv.getCatalog(catalog.name());
-          Assertions.assertTrue(flinkCatalog.isPresent());
-          try {
-            CatalogBaseTable flinkTable =
-                flinkCatalog.get().getTable(new ObjectPath(schemaName, viewName));
-            Assertions.assertEquals(CatalogBaseTable.TableKind.VIEW, flinkTable.getTableKind());
-          } catch (TableNotExistException e) {
-            Assertions.fail("view should exist in Flink catalog: " + e.getMessage());
-          }
-        },
-        true);
+  @Override
+  protected boolean supportViewOperation() {
+    return true;
   }
 
-  @Test
-  public void testListViews() {
-    String schemaName = "test_hive_list_views_db";
-    String tableName = "test_list_view_base";
-    String view1 = "test_list_view_1";
-    String view2 = "test_list_view_2";
-    doWithSchema(
-        currentCatalog(),
-        schemaName,
-        catalog -> {
-          TestUtils.assertTableResult(
-              sql("CREATE TABLE %s (id INT) WITH ('connector'='hive')", tableName),
-              ResultKind.SUCCESS);
-          TestUtils.assertTableResult(
-              sql("CREATE VIEW %s AS SELECT id FROM %s", view1, tableName), ResultKind.SUCCESS);
-          TestUtils.assertTableResult(
-              sql("CREATE VIEW %s AS SELECT id FROM %s", view2, tableName), ResultKind.SUCCESS);
-
-          List<String> views = Arrays.asList(tableEnv.listViews());
-          Assertions.assertTrue(views.contains(view1), "view1 not found in SHOW VIEWS");
-          Assertions.assertTrue(views.contains(view2), "view2 not found in SHOW VIEWS");
-
-          // Tables should not appear in listViews
-          Assertions.assertFalse(views.contains(tableName), "table should not appear in listViews");
-
-          // Verify via Gravitino ViewCatalog
-          ViewCatalog viewCatalog = catalog.asViewCatalog();
-          NameIdentifier[] gravitinoViews = viewCatalog.listViews(Namespace.of(schemaName));
-          List<String> gravitinoViewNames =
-              Arrays.stream(gravitinoViews).map(NameIdentifier::name).collect(Collectors.toList());
-          Assertions.assertTrue(gravitinoViewNames.contains(view1));
-          Assertions.assertTrue(gravitinoViewNames.contains(view2));
-        },
-        true);
-  }
-
-  @Test
-  public void testDropView() {
-    String schemaName = "test_hive_drop_view_db";
-    String tableName = "test_drop_view_base";
-    String viewName = "test_view_drop";
-    doWithSchema(
-        currentCatalog(),
-        schemaName,
-        catalog -> {
-          TestUtils.assertTableResult(
-              sql("CREATE TABLE %s (id INT) WITH ('connector'='hive')", tableName),
-              ResultKind.SUCCESS);
-          TestUtils.assertTableResult(
-              sql("CREATE VIEW %s AS SELECT id FROM %s", viewName, tableName), ResultKind.SUCCESS);
-
-          ViewCatalog viewCatalog = catalog.asViewCatalog();
-          Assertions.assertTrue(
-              viewCatalog.viewExists(NameIdentifier.of(schemaName, viewName)),
-              "view should exist before drop");
-
-          TestUtils.assertTableResult(sql("DROP VIEW %s", viewName), ResultKind.SUCCESS);
-
-          Assertions.assertFalse(
-              viewCatalog.viewExists(NameIdentifier.of(schemaName, viewName)),
-              "view should not exist after drop");
-        },
-        true);
-  }
-
-  @Test
-  public void testAlterViewRename() {
-    String schemaName = "test_hive_rename_view_db";
-    String tableName = "test_rename_view_base";
-    String viewName = "test_view_rename_src";
-    String newViewName = "test_view_rename_dst";
-    doWithSchema(
-        currentCatalog(),
-        schemaName,
-        catalog -> {
-          TestUtils.assertTableResult(
-              sql("CREATE TABLE %s (id INT) WITH ('connector'='hive')", tableName),
-              ResultKind.SUCCESS);
-          TestUtils.assertTableResult(
-              sql("CREATE VIEW %s AS SELECT id FROM %s", viewName, tableName), ResultKind.SUCCESS);
-
-          TestUtils.assertTableResult(
-              sql("ALTER VIEW %s RENAME TO %s", viewName, newViewName), ResultKind.SUCCESS);
-
-          ViewCatalog viewCatalog = catalog.asViewCatalog();
-          Assertions.assertFalse(
-              viewCatalog.viewExists(NameIdentifier.of(schemaName, viewName)),
-              "old view name should not exist");
-          Assertions.assertTrue(
-              viewCatalog.viewExists(NameIdentifier.of(schemaName, newViewName)),
-              "new view name should exist");
-        },
-        true);
-  }
-
-  @Test
-  public void testAlterViewReplaceBody() {
-    String schemaName = "test_hive_replace_view_db";
-    String tableName = "test_replace_view_base";
-    String viewName = "test_view_replace";
-    doWithSchema(
-        currentCatalog(),
-        schemaName,
-        catalog -> {
-          TestUtils.assertTableResult(
-              sql("CREATE TABLE %s (id INT, name STRING) WITH ('connector'='hive')", tableName),
-              ResultKind.SUCCESS);
-          TestUtils.assertTableResult(
-              sql("CREATE VIEW %s AS SELECT id FROM %s", viewName, tableName), ResultKind.SUCCESS);
-
-          // Replace the view body to select a different column
-          TestUtils.assertTableResult(
-              sql("ALTER VIEW %s AS SELECT id, name FROM %s", viewName, tableName),
-              ResultKind.SUCCESS);
-
-          ViewCatalog viewCatalog = catalog.asViewCatalog();
-          View view = viewCatalog.loadView(NameIdentifier.of(schemaName, viewName));
-          Assertions.assertEquals(1, view.representations().length);
-          SQLRepresentation rep = (SQLRepresentation) view.representations()[0];
-          Assertions.assertTrue(
-              rep.sql().contains("name") && rep.sql().contains("id"),
-              "updated view SQL should select both id and name columns");
-        },
-        true);
-  }
-
-  @Test
-  public void testQueryView() {
-    String schemaName = "test_hive_query_view_db";
-    String tableName = "test_query_view_base";
-    String viewName = "test_view_query";
-    doWithSchema(
-        currentCatalog(),
-        schemaName,
-        catalog -> {
-          TestUtils.assertTableResult(
-              sql("CREATE TABLE %s (id INT, name STRING) WITH ('connector'='hive')", tableName),
-              ResultKind.SUCCESS);
-          TestUtils.assertTableResult(
-              sql("CREATE VIEW %s AS SELECT id, name FROM %s WHERE id > 1", viewName, tableName),
-              ResultKind.SUCCESS);
-          TestUtils.assertTableResult(
-              sql("INSERT INTO %s VALUES (1, 'alice'), (2, 'bob'), (3, 'carol')", tableName),
-              ResultKind.SUCCESS_WITH_CONTENT,
-              Row.of(-1L));
-
-          TestUtils.assertTableResult(
-              sql("SELECT * FROM %s ORDER BY id", viewName),
-              ResultKind.SUCCESS_WITH_CONTENT,
-              Row.of(2, "bob"),
-              Row.of(3, "carol"));
-        },
-        true);
-  }
-
-  @Test
-  public void testCreateViewIfNotExists() {
-    String schemaName = "test_hive_create_view_ifte_db";
-    String tableName = "test_create_view_ifte_base";
-    String viewName = "test_view_ifte";
-    doWithSchema(
-        currentCatalog(),
-        schemaName,
-        catalog -> {
-          TestUtils.assertTableResult(
-              sql("CREATE TABLE %s (id INT) WITH ('connector'='hive')", tableName),
-              ResultKind.SUCCESS);
-          TestUtils.assertTableResult(
-              sql("CREATE VIEW %s AS SELECT id FROM %s", viewName, tableName), ResultKind.SUCCESS);
-
-          // Second CREATE VIEW IF NOT EXISTS should succeed without error
-          TestUtils.assertTableResult(
-              sql("CREATE VIEW IF NOT EXISTS %s AS SELECT id FROM %s", viewName, tableName),
-              ResultKind.SUCCESS);
-
-          // View should still exist and be unchanged
-          ViewCatalog viewCatalog = catalog.asViewCatalog();
-          Assertions.assertTrue(viewCatalog.viewExists(NameIdentifier.of(schemaName, viewName)));
-        },
-        true);
-  }
-
-  @Test
-  public void testDropViewIfExists() {
-    String schemaName = "test_hive_drop_view_ifte_db";
-    String tableName = "test_drop_view_ifte_base";
-    String viewName = "test_view_drop_ifte";
-    doWithSchema(
-        currentCatalog(),
-        schemaName,
-        catalog -> {
-          TestUtils.assertTableResult(
-              sql("CREATE TABLE %s (id INT) WITH ('connector'='hive')", tableName),
-              ResultKind.SUCCESS);
-          TestUtils.assertTableResult(
-              sql("CREATE VIEW %s AS SELECT id FROM %s", viewName, tableName), ResultKind.SUCCESS);
-
-          TestUtils.assertTableResult(sql("DROP VIEW %s", viewName), ResultKind.SUCCESS);
-
-          // DROP VIEW IF EXISTS on a non-existent view should not throw
-          TestUtils.assertTableResult(sql("DROP VIEW IF EXISTS %s", viewName), ResultKind.SUCCESS);
-        },
-        true);
-  }
-
-  @Test
-  public void testListTablesDoesNotIncludeViews() {
-    String schemaName = "test_hive_list_tables_no_views_db";
-    String tableName = "test_list_no_view_base";
-    String viewName = "test_list_no_view_view";
-    doWithSchema(
-        currentCatalog(),
-        schemaName,
-        catalog -> {
-          TestUtils.assertTableResult(
-              sql("CREATE TABLE %s (id INT) WITH ('connector'='hive')", tableName),
-              ResultKind.SUCCESS);
-          TestUtils.assertTableResult(
-              sql("CREATE VIEW %s AS SELECT id FROM %s", viewName, tableName), ResultKind.SUCCESS);
-
-          List<String> tables = Arrays.asList(tableEnv.listTables());
-          Assertions.assertTrue(tables.contains(tableName), "table should appear in listTables");
-          Assertions.assertFalse(tables.contains(viewName), "view should not appear in listTables");
-
-          List<String> views = Arrays.asList(tableEnv.listViews());
-          Assertions.assertTrue(views.contains(viewName), "view should appear in listViews");
-          Assertions.assertFalse(views.contains(tableName), "table should not appear in listViews");
-        },
-        true);
+  @Override
+  protected String baseTableConnectorClause() {
+    return " WITH ('connector'='hive')";
   }
 
   private void assertHiveCatalogRead(String databaseName, String tableName, Row expectedRow) {
