+++ flink-connector/flink-common/src/test/java/org/apache/gravitino/flink/connector/integration/test/iceberg/FlinkIcebergCatalogIT.java
@@ -67,6 +67,11 @@ public abstract class FlinkIcebergCatalogIT extends FlinkCommonIT {
     return false;
   }
 
+  @Override
+  protected boolean supportViewOperation() {
+    return true;
+  }
+
   @BeforeAll
   public void before() {
     Preconditions.checkArgument(metalake != null, "metalake should not be null");
