+++ flink-connector/flink-common/src/test/java/org/apache/gravitino/flink/connector/integration/test/paimon/FlinkPaimonHiveBackendIT.java
@@ -56,4 +56,9 @@ public abstract class FlinkPaimonHiveBackendIT extends FlinkPaimonCatalogIT {
   protected String getWarehouse() {
     return warehouse;
   }
+
+  @Override
+  protected boolean supportViewOperation() {
+    return true;
+  }
 }
